refactor: move loki to a separate file
