+++ src/servers/src/http/event.rs
@@ -12,16 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::collections::BTreeMap;
 use std::result::Result as StdResult;
 use std::sync::Arc;
 use std::time::Instant;
 
-use api::v1::value::ValueData;
-use api::v1::{
-    ColumnDataType, ColumnSchema, Row, RowInsertRequest, RowInsertRequests, Rows, SemanticType,
-    Value as GreptimeValue,
-};
+use api::v1::{RowInsertRequest, RowInsertRequests, Rows};
 use axum::body::HttpBody;
 use axum::extract::{FromRequest, Multipart, Path, Query, State};
 use axum::headers::ContentType;
@@ -29,66 +24,34 @@ use axum::http::header::CONTENT_TYPE;
 use axum::http::{Request, StatusCode};
 use axum::response::{IntoResponse, Response};
 use axum::{async_trait, BoxError, Extension, Json, TypedHeader};
-use bytes::Bytes;
 use common_error::ext::ErrorExt;
-use common_query::prelude::GREPTIME_TIMESTAMP;
 use common_query::{Output, OutputData};
 use common_telemetry::{error, warn};
 use datatypes::value::column_data_to_json;
-use hashbrown::HashMap;
-use lazy_static::lazy_static;
-use loki_api::prost_types::Timestamp;
 use pipeline::error::PipelineTransformSnafu;
 use pipeline::util::to_pipeline_version;
 use pipeline::{GreptimeTransformer, PipelineVersion};
-use prost::Message;
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Deserializer, Map, Value};
 use session::context::{Channel, QueryContext, QueryContextRef};
 use snafu::{ensure, OptionExt, ResultExt};
 
 use crate::error::{
-    status_code_to_http_status, CatalogSnafu, DecodeOtlpRequestSnafu, Error, InvalidParameterSnafu,
-    ParseJson5Snafu, ParseJsonSnafu, PipelineSnafu, Result, UnsupportedContentTypeSnafu,
+    status_code_to_http_status, CatalogSnafu, Error, InvalidParameterSnafu, ParseJsonSnafu,
+    PipelineSnafu, Result, UnsupportedContentTypeSnafu,
 };
-use crate::http::extractor::LogTableName;
-use crate::http::header::CONTENT_TYPE_PROTOBUF_STR;
 use crate::http::result::greptime_manage_resp::GreptimedbManageResponse;
 use crate::http::result::greptime_result_v1::GreptimedbV1Response;
 use crate::http::HttpResponse;
 use crate::interceptor::{LogIngestInterceptor, LogIngestInterceptorRef};
 use crate::metrics::{
     METRIC_FAILURE_VALUE, METRIC_HTTP_LOGS_INGESTION_COUNTER, METRIC_HTTP_LOGS_INGESTION_ELAPSED,
-    METRIC_HTTP_LOGS_TRANSFORM_ELAPSED, METRIC_LOKI_LOGS_INGESTION_COUNTER,
-    METRIC_LOKI_LOGS_INGESTION_ELAPSED, METRIC_SUCCESS_VALUE,
+    METRIC_HTTP_LOGS_TRANSFORM_ELAPSED, METRIC_SUCCESS_VALUE,
 };
-use crate::prom_store;
 use crate::query_handler::PipelineHandlerRef;
 
 const GREPTIME_INTERNAL_PIPELINE_NAME_PREFIX: &str = "greptime_";
 const GREPTIME_INTERNAL_IDENTITY_PIPELINE_NAME: &str = "greptime_identity";
-
-const LOKI_TABLE_NAME: &str = "loki_logs";
-const LOKI_LINE_COLUMN: &str = "line";
-
-lazy_static! {
-    static ref LOKI_INIT_SCHEMAS: Vec<ColumnSchema> = vec![
-        ColumnSchema {
-            column_name: GREPTIME_TIMESTAMP.to_string(),
-            datatype: ColumnDataType::TimestampNanosecond.into(),
-            semantic_type: SemanticType::Timestamp.into(),
-            datatype_extension: None,
-            options: None,
-        },
-        ColumnSchema {
-            column_name: LOKI_LINE_COLUMN.to_string(),
-            datatype: ColumnDataType::String.into(),
-            semantic_type: SemanticType::Field.into(),
-            datatype_extension: None,
-            options: None,
-        },
-    ];
-}
 
 #[derive(Debug, Default, Serialize, Deserialize)]
 pub struct LogIngesterQueryParams {
@@ -486,148 +449,6 @@ pub async fn pipeline_dryrun(
 }
 
 #[axum_macros::debug_handler]
-pub async fn loki_ingest(
-    State(log_state): State<LogState>,
-    Extension(mut ctx): Extension<QueryContext>,
-    TypedHeader(content_type): TypedHeader<ContentType>,
-    LogTableName(table_name): LogTableName,
-    bytes: Bytes,
-) -> Result<HttpResponse> {
-    ctx.set_channel(Channel::Loki);
-    let ctx = Arc::new(ctx);
-    let db = ctx.get_db_string();
-    let db_str = db.as_str();
-    let table_name = table_name.unwrap_or_else(|| LOKI_TABLE_NAME.to_string());
-    let exec_timer = Instant::now();
-
-    // decompress req
-    ensure!(
-        content_type.to_string() == CONTENT_TYPE_PROTOBUF_STR,
-        UnsupportedContentTypeSnafu { content_type }
-    );
-    let decompressed = prom_store::snappy_decompress(&bytes).unwrap();
-    let req = loki_api::logproto::PushRequest::decode(&decompressed[..])
-        .context(DecodeOtlpRequestSnafu)?;
-
-    // init schemas
-    let mut schemas = LOKI_INIT_SCHEMAS.clone();
-
-    let mut global_label_key_index: HashMap<String, u16> = HashMap::new();
-    global_label_key_index.insert(GREPTIME_TIMESTAMP.to_string(), 0);
-    global_label_key_index.insert(LOKI_LINE_COLUMN.to_string(), 1);
-
-    let cnt = req.streams.iter().map(|s| s.entries.len()).sum::<usize>();
-    let mut rows = Vec::with_capacity(cnt);
-
-    for stream in req.streams {
-        // parse labels for each row
-        // encoding: https://github.com/grafana/alloy/blob/be34410b9e841cc0c37c153f9550d9086a304bca/internal/component/common/loki/client/batch.go#L114-L145
-        // use very dirty hack to parse labels
-        let labels = stream.labels.replace("=", ":");
-        // use btreemap to keep order
-        let labels: BTreeMap<String, String> = json5::from_str(&labels).context(ParseJson5Snafu)?;
-
-        // process entries
-        for entry in stream.entries {
-            let ts = if let Some(ts) = entry.timestamp {
-                ts
-            } else {
-                continue;
-            };
-            let line = entry.line;
-
-            // create and init row
-            let mut row = Vec::with_capacity(schemas.len());
-            // set ts and line
-            row.push(GreptimeValue {
-                value_data: Some(ValueData::TimestampNanosecondValue(prost_ts_to_nano(&ts))),
-            });
-            row.push(GreptimeValue {
-                value_data: Some(ValueData::StringValue(line)),
-            });
-            for _ in 0..(schemas.len() - 2) {
-                row.push(GreptimeValue { value_data: None });
-            }
-
-            // insert labels
-            for (k, v) in labels.iter() {
-                if let Some(index) = global_label_key_index.get(k) {
-                    // exist in schema
-                    // insert value using index
-                    row[*index as usize] = GreptimeValue {
-                        value_data: Some(ValueData::StringValue(v.clone())),
-                    };
-                } else {
-                    // not exist
-                    // add schema and append to values
-                    schemas.push(ColumnSchema {
-                        column_name: k.clone(),
-                        datatype: ColumnDataType::String.into(),
-                        semantic_type: SemanticType::Tag.into(),
-                        datatype_extension: None,
-                        options: None,
-                    });
-                    global_label_key_index.insert(k.clone(), (schemas.len() - 1) as u16);
-
-                    row.push(GreptimeValue {
-                        value_data: Some(ValueData::StringValue(v.clone())),
-                    });
-                }
-            }
-
-            rows.push(row);
-        }
-    }
-
-    // fill Null for missing values
-    for row in rows.iter_mut() {
-        if row.len() < schemas.len() {
-            for _ in row.len()..schemas.len() {
-                row.push(GreptimeValue { value_data: None });
-            }
-        }
-    }
-
-    let rows = Rows {
-        rows: rows.into_iter().map(|values| Row { values }).collect(),
-        schema: schemas,
-    };
-
-    let ins_req = RowInsertRequest {
-        table_name,
-        rows: Some(rows),
-    };
-    let ins_reqs = RowInsertRequests {
-        inserts: vec![ins_req],
-    };
-
-    let handler = log_state.log_handler;
-    let output = handler.insert(ins_reqs, ctx).await;
-
-    if let Ok(Output {
-        data: OutputData::AffectedRows(rows),
-        meta: _,
-    }) = &output
-    {
-        METRIC_LOKI_LOGS_INGESTION_COUNTER
-            .with_label_values(&[db_str])
-            .inc_by(*rows as u64);
-        METRIC_LOKI_LOGS_INGESTION_ELAPSED
-            .with_label_values(&[db_str, METRIC_SUCCESS_VALUE])
-            .observe(exec_timer.elapsed().as_secs_f64());
-    } else {
-        METRIC_LOKI_LOGS_INGESTION_ELAPSED
-            .with_label_values(&[db_str, METRIC_FAILURE_VALUE])
-            .observe(exec_timer.elapsed().as_secs_f64());
-    }
-
-    let response = GreptimedbV1Response::from_output(vec![output])
-        .await
-        .with_execution_time(exec_timer.elapsed().as_millis() as u64);
-    Ok(response)
-}
-
-#[axum_macros::debug_handler]
 pub async fn log_ingester(
     State(log_state): State<LogState>,
     Query(query_params): Query<LogIngesterQueryParams>,
@@ -811,11 +632,6 @@ pub struct LogState {
     pub ingest_interceptor: Option<LogIngestInterceptorRef<Error>>,
 }
 
-#[inline]
-fn prost_ts_to_nano(ts: &Timestamp) -> i64 {
-    ts.seconds * 1_000_000_000 + ts.nanos as i64
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -849,17 +665,5 @@ mod tests {
         )
         .to_string();
         assert_eq!(a, "[{\"a\":1},{\"b\":2}]");
-    }
-
-    #[test]
-    fn test_ts_to_nano() {
-        // ts = 1731748568804293888
-        // seconds = 1731748568
-        // nano = 804293888
-        let ts = Timestamp {
-            seconds: 1731748568,
-            nanos: 804293888,
-        };
-        assert_eq!(prost_ts_to_nano(&ts), 1731748568804293888);
     }
 }
